refactor: moving component to view directory
